Creates a qq object

An object of class qq includes all the needed information for producing a quantile-quantile plot of p-values. The goal is to seperate the pre-processing of the quantile-quantile plot elements from the graphical rendaring of the object, which could be done using any graphical device including plot_ly and plot in base R.

qqr(x, p = "P", snp, gene, annotation1, annotation2, ...)

Arguments

x
A data.frame which must contain at least the following column:

a p-value, must be numeric

p
A string denoting the column name for the p-values. Default is p = "P". This column must be numeric or integer. Should not have missing, NA, NaN, or NULL values and should be between 0 and 1.

snp
A string denoting the column name for the SNP names (e.g. rs number). More generally, this column could be anything that identifies each point being plotted. For example, in an Epigenomewide association study (EWAS) this could be the probe name or cg number. This column should be a character. This argument is optional, however it is necessary to specify if you want to highlight points on the plot using the highlight argument in the qqly function

gene
A string denoting the column name for the GENE names. This column could be a character or numeric. More generally this could be any annotation information that you want to include in the plot. This argument is optional.

annotation1
A string denoting the column name for an annotation. This column could be a character or numeric. This could be any annotation information that you want to include in the plot (e.g. zscore, effect size, minor allele frequency). This argument is optional.

annotation2
A string denoting the column name for an annotation. This column could be a character or numeric. This could be any annotation information that you want to include in the plot (e.g. zscore, effect size, minor allele frequency). This argument is optional.

...
currently ignored

Source

The calculation of the expected p-value is taken from the qq function from the qqman package

Value

An list object of class qqr with the following elements

data

processed data to be used for plotting the Q-Q plot including the observed and expected p-values on the -log10 scale

pName, snpName, geneName, annotation1Name, annotation2Name

The names of the columns corresponding to the data provided. This information is used for annotating the plot in the qqly function

Note

This function will return an error if any of the p-values are NA, less than 0 or greater than 1

See also

Examples

library(manhattanly)
qqrObj <- qqr(HapMap, snp = "SNP", highlight = significantSNP)
class(qqrObj)

#> [1] "qqr"

head(qqrObj[["data"]])

#>                P        SNP OBSERVED EXPECTED
#> 4346 6.75010e-10 rs16900007 9.170690 4.459754
#> 4347 3.41101e-09  rs1039325 8.467117 3.982633
#> 4344 3.95101e-09  rs2195387 8.403292 3.760784
#> 4338 4.70701e-09 rs16898108 8.327255 3.614656
#> 4342 5.02201e-09  rs6450718 8.299122 3.505512
#> 4341 6.22801e-09 rs16899251 8.205651 3.418362